% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beta_Generation.R
\name{Beta_Generation}
\alias{Beta_Generation}
\title{True Beta Generation}
\usage{
Beta_Generation(P, k, type = 1)
}
\arguments{
\item{P}{Numeric. The number of predictors in the model.}

\item{k}{Numeric. The number of non-zero coefficients in the true coefficient vector.}

\item{type}{Numeric (optional). A parameter to specify the structure of the true coefficient vector.
Possible values:
\itemize{
  \item 1: Sparse coefficients equal to 1 at roughly equally spaced indices.
  \item 2: First \( k \) components equal to 1.
  \item 3: First \( k \) components with values equally spaced between 10 and 0.5.
  \item 5: First \( k \) components equal to 1, then exponentially decaying coefficients.
}}
}
\value{
Numeric vector. The generated true coefficient vector.
}
\description{
This function generates a true coefficient vector based on the specified type.
The vector can represent different patterns of sparsity, such as coefficients that are equal to 1 at equally spaced indices,
or decaying exponentially, among other types.
}
\examples{
# Example of generating a beta vector with 50 predictors and 10 non-zero coefficients
beta <- generate_beta(P = 50, k = 10, type = 1)
print(beta)

}
