% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Generation.R
\name{Data_Generation}
\alias{Data_Generation}
\title{Data Generation for Linear Model}
\usage{
Data_Generation(N, P, family, SNR, rho, beta)
}
\arguments{
\item{N}{Numeric. The number of observations.}

\item{P}{Numeric. The number of predictors.}

\item{family}{Character string specifying the distribution of the outcome variable.
One of "Binomial", "Bernoulli", "Normal", "Poisson", "Gamma", or "Exponential".}

\item{SNR}{Numeric The signal-to-noise ratio. Controls the variance of the outcome variable.}

\item{rho}{Numeric The correlation coefficient between predictors in the design matrix.}

\item{beta}{Numeric vector. The true coefficient vector used to generate the outcome variable.}
}
\value{
A list containing the following elements:
  \item{X}{Numeric matrix. The design matrix.}
  \item{Y}{Numeric vector. The outcome vector.}
  \item{cov}{Numeric matrix. The covariance matrix of the predictors.}
  \item{variance}{Numeric. The variance of the outcome variable.}
}
\description{
This function generates synthetic dataset for linear model and Generalized linear model.
}
\details{
The design matrix is generated from a multivariate normal distribution with mean zero and a specified correlation structure.
The true coefficient vector  is used to compute the linear predictor, and the outcome vector is generated
as a linear combination of the predictors with added noise, controlled by the signal-to-noise ratio (SNR).
}
\examples{
# Example of generating a dataset with 100 observations and 50 predictors
result <- Generation(N = 100, P = 50, family = "Normal", SNR = 10, rho = 0.5, beta = rep(1, 20))
X <- result[[1]]
Y <- result[[2]]

}
