% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{install_julia_dependencies}
\alias{install_julia_dependencies}
\title{Install Julia and Required Packages}
\usage{
install_julia_dependencies()
}
\description{
Installs Julia and required Julia packages (e.g., 'Distributions', 'LinearAlgebra').
Only call this function manually.
}
