% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPtoY.R
\name{LP_to_Y}
\alias{LP_to_Y}
\title{Transform Linear Predictor to Outcome Y}
\usage{
LP_to_Y(
  X,
  true_beta,
  family = "Normal",
  n_trials = NULL,
  std = NULL,
  shape = NULL,
  cov_matrix = NULL,
  n_categories = NULL
)
}
\arguments{
\item{X}{Design matrix}

\item{true_beta}{Coefficient vector/matrix}

\item{family}{Distribution family ("Normal", "Binomial", etc.)}

\item{n_trials}{Number of trials for Binomial family}

\item{std}{Standard deviation for Normal family}

\item{shape}{Shape parameter for Gamma family}

\item{cov_matrix}{Covariance matrix for MultivariateNormal}

\item{n_categories}{Number of categories for Multinomial}
}
\value{
Simulated outcome values
}
\description{
Converts linear predictors to outcomes using Julia implementation
}
