% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty_selection.R
\name{information_criteria}
\alias{information_criteria}
\alias{Calculate_AIC}
\alias{Calculate_AICc}
\alias{Calculate_AttIC}
\alias{Calculate_SIC}
\alias{Calculate_BIC}
\alias{Calculate_CAIC}
\alias{Calculate_CAICF}
\alias{Calculate_GIC2}
\alias{Calculate_GIC3}
\alias{Calculate_GIC4}
\alias{Calculate_GIC5}
\alias{Calculate_GIC6}
\alias{Calculate_AIC_Short}
\alias{Calculate_AICc_Short}
\alias{Calculate_AttIC_Short}
\alias{Calculate_SIC_Short}
\alias{Calculate_BIC_Short}
\alias{Calculate_CAIC_Short}
\alias{Calculate_CAICF_Short}
\alias{Calculate_GIC2_Short}
\alias{Calculate_GIC3_Short}
\alias{Calculate_GIC4_Short}
\alias{Calculate_GIC5_Short}
\alias{Calculate_GIC6_Short}
\title{Information Criteria Calculations via Julia}
\usage{
Calculate_AIC(Y, X)

Calculate_AICc(Y, X)

Calculate_AttIC(Y, X)

Calculate_SIC(Y, X)

Calculate_BIC(Y, X)

Calculate_CAIC(Y, X)

Calculate_CAICF(Y, X)

Calculate_GIC2(Y, X)

Calculate_GIC3(Y, X)

Calculate_GIC4(Y, X)

Calculate_GIC5(Y, X)

Calculate_GIC6(Y, X)

Calculate_AIC_Short(Y, X, Inverse)

Calculate_AICc_Short(Y, X, Inverse)

Calculate_AttIC_Short(Y, X, Inverse)

Calculate_SIC_Short(Y, X, Inverse)

Calculate_BIC_Short(Y, X, Inverse)

Calculate_CAIC_Short(Y, X, Inverse)

Calculate_CAICF_Short(Y, X, Inverse)

Calculate_GIC2_Short(Y, X, Inverse)

Calculate_GIC3_Short(Y, X, Inverse)

Calculate_GIC4_Short(Y, X, Inverse)

Calculate_GIC5_Short(Y, X, Inverse)

Calculate_GIC6_Short(Y, X, Inverse)
}
\arguments{
\item{Y}{Outcome vector or matrix (numeric).}

\item{X}{Design matrix (numeric).}
}
\value{
A list containing:
\itemize{
  \item The calculated criterion value
  \item The inverse matrix (X'X)^{-1}
}
}
\description{
These functions calculate various information criteria using optimized Julia implementations.
Each function returns a list containing the criterion value and inverse matrix.
}
