% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty_selection.R
\name{information_criteria}
\alias{information_criteria}
\alias{calculate_aic}
\alias{calculate_aicc}
\alias{calculate_att}
\alias{calculate_sic}
\alias{calculate_bic}
\alias{calculate_caic}
\alias{calculate_caicf}
\alias{calculate_gic2}
\alias{calculate_gic3}
\alias{calculate_gic4}
\alias{calculate_gic5}
\alias{calculate_gic6}
\alias{calculate_aic_short}
\alias{calculate_aicc_short}
\alias{calculate_att_short}
\alias{calculate_sic_short}
\alias{calculate_bic_short}
\alias{calculate_caic_short}
\alias{calculate_caicf_short}
\alias{calculate_gic2_short}
\alias{calculate_gic3_short}
\alias{calculate_gic4_short}
\alias{calculate_gic5_short}
\alias{calculate_gic6_short}
\title{Information Criteria Calculations via Julia}
\usage{
calculate_aic(Y, X)

calculate_aicc(Y, X)

calculate_att(Y, X)

calculate_sic(Y, X)

calculate_bic(Y, X)

calculate_caic(Y, X)

calculate_caicf(Y, X)

calculate_gic2(Y, X)

calculate_gic3(Y, X)

calculate_gic4(Y, X)

calculate_gic5(Y, X)

calculate_gic6(Y, X)

calculate_aic_short(Y, X, Inverse)

calculate_aicc_short(Y, X, Inverse)

calculate_att_short(Y, X, Inverse)

calculate_sic_short(Y, X, Inverse)

calculate_bic_short(Y, X, Inverse)

calculate_caic_short(Y, X, Inverse)

calculate_caicf_short(Y, X, Inverse)

calculate_gic2_short(Y, X, Inverse)

calculate_gic3_short(Y, X, Inverse)

calculate_gic4_short(Y, X, Inverse)

calculate_gic5_short(Y, X, Inverse)

calculate_gic6_short(Y, X, Inverse)
}
\arguments{
\item{Y}{Outcome vector or matrix (numeric).}

\item{X}{Design matrix (numeric).}
}
\value{
A list containing:
\itemize{
  \item The calculated criterion value
  \item The inverse matrix (X'X)^{-1}
}
}
\description{
These functions calculate various information criteria using optimized Julia implementations.
Each function returns a list containing the criterion value and inverse matrix.
}
